handle 'date' and 'main' lookups
The logic for these lookups is simple and lossy, but that's acceptable
since we can at least report that these lookups will be made. For some
users it is reasonable to simply reject a log line at the presence of
`main` or `date`, regardless of later findings. For other users, it
doesn't seem possible to have a general transformation between a `main`
or `date` lookup, and a string useful to build other more concerning
lookup tokens (like `jndi` or `env`).